Refactor FAISS toolkit scripts and update documentation
- Merged `vector_faiss_load.py` and `vector_faiss_write.py` into a single script `vector_faiss.py` to streamline FAISS index operations.
- Updated the CLI tool to support both 'read' and 'save' operations.
- Modified `main()` function to handle the new 'operation' argument, allowing users to specify 'read' or 'save'.
- Updated `examples/rag/chat.sh` to use the new `cllm-vector-faiss` script for both reading and writing operations.
- Deleted old documentation files `cllm-vector-faiss-load.md` and `cllm-vector-faiss-write.md`.
- Created a new consolidated documentation file `cllm-vector-faiss.md` to reflect the changes and provide usage examples for the updated CLI tool.
- Updated `pyproject.toml` to reflect the new script name and entry point for `cllm-vector-faiss`.
- Adjusted logging configuration and environment variable handling for better consistency and error handling.

This refactor simplifies the FAISS toolkit, making it more user-friendly and easier to maintain.

closes #10
