--- docs/toolkit/cllm-vector-faiss-load.md
@@ -1,31 +1,0 @@
-# FAISS Index Search Tool
-
-This script provides a command-line interface (CLI) for performing similarity searches on a FAISS index using OpenAI embeddings. Below is a detailed explanation of the script and its components.
-
-## Script Overview
-
-The script is designed to load a FAISS index, perform a similarity search based on a query, and return the results in JSON format. It uses environment variables for configuration and supports logging for error handling and debugging.
-
-## Usage
-
-To use the script, run it from the command line with the appropriate arguments:
-
-```sh
-cllm-vector-faiss-load -i <index_name> -m <model_name> -k <number_of_results> <query>
-```
-
-### Example
-
-```sh
-cllm-vector-faiss-load -i my_faiss_index -m text-embedding-3-small -k 5 "What is the capital of France?"
-```
-
-This command will search the FAISS index named `my_faiss_index` using the `text-embedding-3-small` model for the query "What is the capital of France?" and return the top 5 results.
-
-## Logging
-
-The script uses Python's built-in logging module to log errors and other information. The log level can be set using the `CLLM_LOGLEVEL` environment variable.
-
-## Conclusion
-
-This script provides a robust and flexible tool for performing similarity searches on FAISS indexes using OpenAI embeddings. It is designed to be easy to use and extend, making it a valuable tool for various information retrieval tasks.
