--- docs/toolkit/cllm-vector-faiss-write.md
@@ -1,38 +1,0 @@
-# FAISS Indexing Script
-
-This script is designed to save data to a FAISS index using OpenAI embeddings. Below is a detailed explanation of the script, its components, and how to use it.
-
-## Script Overview
-
-The script performs the following tasks:
-1. Configures logging.
-2. Loads configuration from environment variables or uses default values.
-3. Defines a function to save data to a FAISS index.
-4. Provides a command-line interface (CLI) for users to interact with the script.
-
-### Configuration
-
-The script sets the directory for saving the FAISS index, defaulting to a `.cllm` directory in the current working directory if not specified in the environment variables.
-
-```python
-CLLM_DIR = os.getenv("CLLM_DIR", f"{os.getcwd()}/.cllm")
-```
-
-## Usage
-
-To use the script, you can run it from the command line with the following options:
-
-- `-i` or `--index`: Path to save the FAISS index (default: `faiss_index`).
-- `-m` or `--model`: Name of the OpenAI embedding model to use (default: `text-embedding-3-small`).
-
-Example command:
-
-```sh
-python cllm-vector-faiss-write -i my_faiss_index -m text-embedding-3-small < data.json
-```
-
-This command reads JSON data from `data.json` and saves it to a FAISS index named `my_faiss_index` using the `text-embedding-3-small` model.
-
-## Logging
-
-The script uses Python's built-in logging module to log errors and other information. The log level can be set using the `CLLM_LOGLEVEL` environment variable.
