+++ pyproject.toml
@@ -28,12 +28,11 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
 cllm = 'cllm.cllm:main'
-cllm-splitter-text = 'cllm.toolkit.cllm_splitter_text:main'
+cllm-splitter-text = 'cllm.toolkit.splitter_text:main'
 cllm-scraper = 'cllm.toolkit.scraper:main'
 cllm-search-google = 'cllm.toolkit.search_google:main'
 cllm-load-files = 'cllm.toolkit.load_files:main'
-cllm-vector-faiss-load = 'cllm.toolkit.vector_faiss_load:main'
-cllm-vector-faiss-write = 'cllm.toolkit.vector_faiss_write:main'
+cllm-vector-faiss = 'cllm.toolkit.vector_faiss:main'
 cllm-splitter-docs = 'cllm.toolkit.splitter_docs:main'
 cllm-load-dir = 'cllm.toolkit.load_dir:main'
 cllm-gen-dalle = 'cllm.toolkit.gen_dalle:main'
